enroll_api: make sure stdout can be parsed as valid JSON
The 'json()' method of the 'response' class (in the 'requests' python
package) returns a Python dict structure with the JSON data in it. It
isn't a JSON class per se, so passing it to 'print()' causes it to render
as a python struct, which isn't valid JSON. Use json.dumps() instead to
render it. Also, remove other possible outputs so that the stdout can be
passed directly to 'jq' and other JSON-consuming tools.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>