build: consolidate package installation
Move all the debian package installation from individual app containers
into the "base platform", so that output images are only differentiated
by their HCP details. This penalizes the swtpmsvc and client image sizes
but results in greater overall efficiency. It's especially helpful with
respect to rebuild time when modifying HCP specifics - as no package
installs are in the path of the rebuilding.

This also means that 'git' is in the base platform, so the wrapper script
in 'builder' is no longer needed.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>