build: remove HCP_{DSPACE,DTAG}, use defaults.env
Since we changed the run-time (and testing) support to use docker-compose,
the build is no longer creating docker networks or volumes, and its only
creation of containers is for running build steps, which uses ephemeral
naming. So the only naming that the build performs is in the image
generation. The docker-compose source of truth (defaults.env) also needs
to know the image naming, and has HCP_IMAGE_PREFIX and HCP_IMAGE_TAG
defined for this purpose. As such, we remove HCP_{DSPACE,DTAG}, and now
have the build source HCP_IMAGE_{PREFIX,TAG} from defaults.env instead.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>